Split up some files into modules/impl FS functions
