+++ c-scape/src/lib.rs
@@ -9,6 +9,7 @@
 #![feature(strict_provenance)]
 #![deny(fuzzy_provenance_casts)]
 #![deny(lossy_provenance_casts)]
+#![feature(try_blocks)]
 
 extern crate alloc;
 extern crate compiler_builtins;
@@ -45,7 +46,7 @@ use core::ffi::c_void;
 #[cfg(not(target_arch = "riscv64"))]
 use core::mem::align_of;
 #[cfg(not(target_os = "wasi"))]
-use core::mem::{size_of, transmute, zeroed};
+use core::mem::{size_of, zeroed};
 use core::ptr::{self, null, null_mut};
 use core::slice;
 #[cfg(feature = "threads")]
@@ -55,15 +56,13 @@ use core::sync::atomic::{AtomicBool, AtomicU32};
 use errno::{set_errno, Errno};
 use error_str::error_str;
 use libc::{c_char, c_int, c_long, c_uint, c_ulong};
-#[cfg(not(target_os = "wasi"))]
-use memoffset::offset_of;
 #[cfg(feature = "threads")]
 use parking_lot::lock_api::{RawMutex as _, RawRwLock};
 #[cfg(feature = "threads")]
 use parking_lot::RawMutex;
-use rustix::fd::{AsFd, AsRawFd, BorrowedFd, FromRawFd, IntoRawFd};
+use rustix::fd::{BorrowedFd, FromRawFd, IntoRawFd};
 use rustix::ffi::ZStr;
-use rustix::fs::{cwd, AtFlags, FdFlags, Mode, OFlags};
+use rustix::fs::{AtFlags, Mode, OFlags};
 #[cfg(any(target_os = "android", target_os = "linux"))]
 use rustix::io::EventfdFlags;
 use rustix::io::OwnedFd;
@@ -81,6 +80,9 @@ use {
     rustix::net::{IpAddr, SocketAddrV4, SocketAddrV6},
     sync_resolve::resolve_host,
 };
+
+// fs
+mod fs;
 
 // errno
 
@@ -106,681 +108,6 @@ unsafe extern "C" fn __xpg_strerror_r(errnum: c_int, buf: *mut c_char, buflen: u
     out.copy_from_slice(message.as_bytes());
     out[out.len() - 1] = b'\0';
     0
-}
-
-// fs
-
-#[no_mangle]
-unsafe extern "C" fn openat(
-    fd: c_int,
-    pathname: *const c_char,
-    flags: c_int,
-    mode: libc::mode_t,
-) -> c_int {
-    libc!(libc::openat(fd, pathname, flags, mode));
-
-    let fd = BorrowedFd::borrow_raw(fd);
-    let flags = OFlags::from_bits(flags as _).unwrap();
-    let mode = if flags.contains(OFlags::CREATE) {
-        Mode::from_bits((mode & !libc::S_IFMT) as _).unwrap()
-    } else {
-        Mode::empty()
-    };
-    match convert_res(rustix::fs::openat(
-        &fd,
-        ZStr::from_ptr(pathname.cast()),
-        flags,
-        mode,
-    )) {
-        Some(fd) => fd.into_raw_fd(),
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn open64(pathname: *const c_char, flags: c_int, mode: libc::mode_t) -> c_int {
-    libc!(libc::open64(pathname, flags, mode));
-
-    let flags = OFlags::from_bits(flags as _).unwrap();
-    let mode = Mode::from_bits((mode & !libc::S_IFMT) as _).unwrap();
-    match convert_res(rustix::fs::openat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        flags,
-        mode,
-    )) {
-        Some(fd) => fd.into_raw_fd(),
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn open() {
-    //libc!(libc::open())
-    unimplemented!("open")
-}
-
-#[no_mangle]
-unsafe extern "C" fn readlink(pathname: *const c_char, buf: *mut c_char, bufsiz: usize) -> isize {
-    libc!(libc::readlink(pathname, buf, bufsiz));
-
-    let path = match convert_res(rustix::fs::readlinkat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        Vec::new(),
-    )) {
-        Some(path) => path,
-        None => return -1,
-    };
-    let bytes = path.as_bytes();
-    let min = core::cmp::min(bytes.len(), bufsiz);
-    slice::from_raw_parts_mut(buf.cast(), min).copy_from_slice(bytes);
-    min as isize
-}
-
-#[no_mangle]
-unsafe extern "C" fn stat() -> c_int {
-    //libc!(libc::stat())
-    unimplemented!("stat")
-}
-
-#[no_mangle]
-unsafe extern "C" fn stat64(pathname: *const c_char, stat_: *mut rustix::fs::Stat) -> c_int {
-    libc!(libc::stat64(pathname, checked_cast!(stat_)));
-
-    match convert_res(rustix::fs::statat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        AtFlags::empty(),
-    )) {
-        Some(r) => {
-            *stat_ = r;
-            0
-        }
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn fstat(_fd: c_int, _stat: *mut libc::stat) -> c_int {
-    libc!(libc::fstat(_fd, _stat));
-    unimplemented!("fstat")
-}
-
-#[no_mangle]
-unsafe extern "C" fn fstat64(fd: c_int, stat_: *mut rustix::fs::Stat) -> c_int {
-    libc!(libc::fstat64(fd, checked_cast!(stat_)));
-
-    match convert_res(rustix::fs::fstat(&BorrowedFd::borrow_raw(fd))) {
-        Some(r) => {
-            *stat_ = r;
-            0
-        }
-        None => -1,
-    }
-}
-
-#[cfg(any(target_os = "android", target_os = "linux"))]
-#[no_mangle]
-unsafe extern "C" fn statx(
-    dirfd_: c_int,
-    path: *const c_char,
-    flags: c_int,
-    mask: c_uint,
-    stat_: *mut rustix::fs::Statx,
-) -> c_int {
-    libc!(libc::statx(dirfd_, path, flags, mask, checked_cast!(stat_)));
-
-    if path.is_null() || stat_.is_null() {
-        set_errno(Errno(libc::EFAULT));
-        return -1;
-    }
-
-    let flags = AtFlags::from_bits(flags as _).unwrap();
-    let mask = rustix::fs::StatxFlags::from_bits(mask).unwrap();
-    match convert_res(rustix::fs::statx(
-        &BorrowedFd::borrow_raw(dirfd_),
-        ZStr::from_ptr(path.cast()),
-        flags,
-        mask,
-    )) {
-        Some(r) => {
-            *stat_ = r;
-            0
-        }
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn realpath(path: *const c_char, resolved_path: *mut c_char) -> *mut c_char {
-    libc!(libc::realpath(path, resolved_path));
-
-    let mut buf = [0; libc::PATH_MAX as usize];
-    match realpath_ext::realpath_raw(
-        ZStr::from_ptr(path.cast()).to_bytes(),
-        &mut buf,
-        realpath_ext::RealpathFlags::empty(),
-    ) {
-        Ok(len) => {
-            if resolved_path.is_null() {
-                let ptr = malloc(len + 1).cast::<u8>();
-                if ptr.is_null() {
-                    set_errno(Errno(libc::ENOMEM));
-                    return null_mut();
-                }
-                slice::from_raw_parts_mut(ptr, len).copy_from_slice(&buf[..len]);
-                *ptr.add(len) = b'\0';
-                ptr.cast::<c_char>()
-            } else {
-                memcpy(
-                    resolved_path.cast::<c_void>(),
-                    buf[..len].as_ptr().cast::<c_void>(),
-                    len,
-                );
-                *resolved_path.add(len) = b'\0' as _;
-                resolved_path
-            }
-        }
-        Err(err) => {
-            set_errno(Errno(err));
-            null_mut()
-        }
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn fcntl(fd: c_int, cmd: c_int, mut args: ...) -> c_int {
-    match cmd {
-        libc::F_GETFL => {
-            libc!(libc::fcntl(fd, libc::F_GETFL));
-            let fd = BorrowedFd::borrow_raw(fd);
-            match convert_res(rustix::fs::fcntl_getfl(&fd)) {
-                Some(flags) => flags.bits() as _,
-                None => -1,
-            }
-        }
-        libc::F_SETFD => {
-            let flags = args.arg::<c_int>();
-            libc!(libc::fcntl(fd, libc::F_SETFD, flags));
-            let fd = BorrowedFd::borrow_raw(fd);
-            match convert_res(rustix::fs::fcntl_setfd(
-                &fd,
-                FdFlags::from_bits(flags as _).unwrap(),
-            )) {
-                Some(()) => 0,
-                None => -1,
-            }
-        }
-        #[cfg(not(target_os = "wasi"))]
-        libc::F_DUPFD_CLOEXEC => {
-            let arg = args.arg::<c_int>();
-            libc!(libc::fcntl(fd, libc::F_DUPFD_CLOEXEC, arg));
-            let fd = BorrowedFd::borrow_raw(fd);
-            match convert_res(rustix::fs::fcntl_dupfd_cloexec(&fd, arg)) {
-                Some(fd) => fd.into_raw_fd(),
-                None => -1,
-            }
-        }
-        _ => panic!("unrecognized fnctl({})", cmd),
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn mkdir(pathname: *const c_char, mode: libc::mode_t) -> c_int {
-    libc!(libc::mkdir(pathname, mode));
-
-    let mode = Mode::from_bits((mode & !libc::S_IFMT) as _).unwrap();
-    match convert_res(rustix::fs::mkdirat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        mode,
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn fdatasync(fd: c_int) -> c_int {
-    libc!(libc::fdatasync(fd));
-
-    match convert_res(rustix::fs::fdatasync(&BorrowedFd::borrow_raw(fd))) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn fstatat64(
-    fd: c_int,
-    pathname: *const c_char,
-    stat_: *mut rustix::fs::Stat,
-    flags: c_int,
-) -> c_int {
-    libc!(libc::fstatat64(fd, pathname, checked_cast!(stat_), flags));
-
-    let flags = AtFlags::from_bits(flags as _).unwrap();
-    match convert_res(rustix::fs::statat(
-        &BorrowedFd::borrow_raw(fd),
-        ZStr::from_ptr(pathname.cast()),
-        flags,
-    )) {
-        Some(r) => {
-            *stat_ = r;
-            0
-        }
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn fsync(fd: c_int) -> c_int {
-    libc!(libc::fsync(fd));
-
-    match convert_res(rustix::fs::fdatasync(&BorrowedFd::borrow_raw(fd))) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn ftruncate64(fd: c_int, length: i64) -> c_int {
-    libc!(libc::ftruncate64(fd, length));
-
-    match convert_res(rustix::fs::ftruncate(
-        &BorrowedFd::borrow_raw(fd),
-        length as u64,
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn rename(old: *const c_char, new: *const c_char) -> c_int {
-    libc!(libc::rename(old, new));
-
-    match convert_res(rustix::fs::renameat(
-        &cwd(),
-        ZStr::from_ptr(old.cast()),
-        &cwd(),
-        ZStr::from_ptr(new.cast()),
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn rmdir(pathname: *const c_char) -> c_int {
-    libc!(libc::rmdir(pathname));
-
-    match convert_res(rustix::fs::unlinkat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        AtFlags::REMOVEDIR,
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn unlinkat(fd: c_int, pathname: *const c_char, flags: c_int) -> c_int {
-    libc!(libc::unlinkat(fd, pathname, flags));
-
-    let fd = BorrowedFd::borrow_raw(fd);
-    let flags = AtFlags::from_bits(flags as _).unwrap();
-    match convert_res(rustix::fs::unlinkat(
-        &fd,
-        ZStr::from_ptr(pathname.cast()),
-        flags,
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn unlink(pathname: *const c_char) -> c_int {
-    libc!(libc::unlink(pathname));
-
-    match convert_res(rustix::fs::unlinkat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        AtFlags::empty(),
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn lseek64(fd: c_int, offset: i64, whence: c_int) -> i64 {
-    libc!(libc::lseek64(fd, offset, whence));
-
-    let seek_from = match whence {
-        libc::SEEK_SET => rustix::io::SeekFrom::Start(offset as u64),
-        libc::SEEK_CUR => rustix::io::SeekFrom::Current(offset),
-        libc::SEEK_END => rustix::io::SeekFrom::End(offset),
-        _ => panic!("unrecognized whence({})", whence),
-    };
-    match convert_res(rustix::fs::seek(&BorrowedFd::borrow_raw(fd), seek_from)) {
-        Some(offset) => offset as i64,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn lstat64(pathname: *const c_char, stat_: *mut rustix::fs::Stat) -> c_int {
-    libc!(libc::lstat64(pathname, checked_cast!(stat_)));
-
-    match convert_res(rustix::fs::statat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        AtFlags::SYMLINK_NOFOLLOW,
-    )) {
-        Some(r) => {
-            *stat_ = r;
-            0
-        }
-        None => -1,
-    }
-}
-
-struct MustangDir {
-    dir: rustix::fs::Dir,
-    dirent: libc::dirent64,
-}
-
-#[no_mangle]
-unsafe extern "C" fn opendir(pathname: *const c_char) -> *mut c_void {
-    libc!(libc::opendir(pathname).cast());
-
-    match convert_res(rustix::fs::openat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        OFlags::RDONLY | OFlags::DIRECTORY | OFlags::CLOEXEC,
-        Mode::empty(),
-    )) {
-        Some(fd) => fdopendir(fd.into_raw_fd()),
-        None => null_mut(),
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn fdopendir(fd: c_int) -> *mut c_void {
-    libc!(libc::fdopendir(fd).cast());
-
-    match convert_res(rustix::fs::Dir::from(OwnedFd::from_raw_fd(fd))) {
-        Some(dir) => Box::into_raw(Box::new(MustangDir {
-            dir,
-            dirent: zeroed(),
-        }))
-        .cast(),
-        None => null_mut(),
-    }
-}
-
-#[cfg(not(target_os = "wasi"))]
-#[no_mangle]
-unsafe extern "C" fn readdir64_r(
-    dir: *mut c_void,
-    entry: *mut libc::dirent64,
-    ptr: *mut *mut libc::dirent64,
-) -> c_int {
-    libc!(libc::readdir64_r(
-        dir.cast(),
-        checked_cast!(entry),
-        checked_cast!(ptr)
-    ));
-
-    let mustang_dir = dir.cast::<MustangDir>();
-    let dir = &mut (*mustang_dir).dir;
-    match dir.read() {
-        None => {
-            *ptr = null_mut();
-            0
-        }
-        Some(Ok(e)) => {
-            let file_type = match e.file_type() {
-                rustix::fs::FileType::RegularFile => libc::DT_REG,
-                rustix::fs::FileType::Directory => libc::DT_DIR,
-                rustix::fs::FileType::Symlink => libc::DT_LNK,
-                rustix::fs::FileType::Fifo => libc::DT_FIFO,
-                rustix::fs::FileType::Socket => libc::DT_SOCK,
-                rustix::fs::FileType::CharacterDevice => libc::DT_CHR,
-                rustix::fs::FileType::BlockDevice => libc::DT_BLK,
-                rustix::fs::FileType::Unknown => libc::DT_UNKNOWN,
-            };
-            *entry = libc::dirent64 {
-                d_ino: e.ino(),
-                d_off: 0, // We don't implement `seekdir` yet anyway.
-                d_reclen: (offset_of!(libc::dirent64, d_name) + e.file_name().to_bytes().len() + 1)
-                    .try_into()
-                    .unwrap(),
-                d_type: file_type,
-                d_name: [0; 256],
-            };
-            let len = core::cmp::min(256, e.file_name().to_bytes().len());
-            (*entry).d_name[..len].copy_from_slice(transmute(e.file_name().to_bytes()));
-            *ptr = entry;
-            0
-        }
-        Some(Err(err)) => err.raw_os_error(),
-    }
-}
-
-#[cfg(not(target_os = "wasi"))]
-#[no_mangle]
-unsafe extern "C" fn readdir64(dir: *mut c_void) -> *mut libc::dirent64 {
-    libc!(libc::readdir64(dir.cast(),));
-
-    let mustang_dir = dir.cast::<MustangDir>();
-    let dir = &mut (*mustang_dir).dir;
-    match dir.read() {
-        None => null_mut(),
-        Some(Ok(e)) => {
-            let file_type = match e.file_type() {
-                rustix::fs::FileType::RegularFile => libc::DT_REG,
-                rustix::fs::FileType::Directory => libc::DT_DIR,
-                rustix::fs::FileType::Symlink => libc::DT_LNK,
-                rustix::fs::FileType::Fifo => libc::DT_FIFO,
-                rustix::fs::FileType::Socket => libc::DT_SOCK,
-                rustix::fs::FileType::CharacterDevice => libc::DT_CHR,
-                rustix::fs::FileType::BlockDevice => libc::DT_BLK,
-                rustix::fs::FileType::Unknown => libc::DT_UNKNOWN,
-            };
-            (*mustang_dir).dirent = libc::dirent64 {
-                d_ino: e.ino(),
-                d_off: 0, // We don't implement `seekdir` yet anyway.
-                d_reclen: (offset_of!(libc::dirent64, d_name) + e.file_name().to_bytes().len() + 1)
-                    .try_into()
-                    .unwrap(),
-                d_type: file_type,
-                d_name: [0; 256],
-            };
-            let len = core::cmp::min(256, e.file_name().to_bytes().len());
-            (*mustang_dir).dirent.d_name[..len]
-                .copy_from_slice(transmute(e.file_name().to_bytes()));
-            &mut (*mustang_dir).dirent
-        }
-        Some(Err(err)) => {
-            set_errno(Errno(err.raw_os_error()));
-            null_mut()
-        }
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn closedir(dir: *mut c_void) -> c_int {
-    libc!(libc::closedir(dir.cast()));
-
-    drop(Box::<MustangDir>::from_raw(dir.cast()));
-    0
-}
-
-#[no_mangle]
-unsafe extern "C" fn dirfd(dir: *mut c_void) -> c_int {
-    libc!(libc::dirfd(dir.cast()));
-
-    let dir = dir.cast::<MustangDir>();
-    (*dir).dir.as_fd().as_raw_fd()
-}
-
-#[no_mangle]
-unsafe extern "C" fn readdir() {
-    //libc!(libc::readdir());
-    unimplemented!("readdir")
-}
-
-#[no_mangle]
-unsafe extern "C" fn rewinddir() {
-    //libc!(libc::rewinddir());
-    unimplemented!("rewinddir")
-}
-
-#[no_mangle]
-unsafe extern "C" fn scandir() {
-    //libc!(libc::scandir());
-    unimplemented!("scandir")
-}
-
-#[no_mangle]
-unsafe extern "C" fn seekdir() {
-    //libc!(libc::seekdir());
-    unimplemented!("seekdir")
-}
-
-#[no_mangle]
-unsafe extern "C" fn telldir() {
-    //libc!(libc::telldir());
-    unimplemented!("telldir")
-}
-
-#[cfg(any(target_os = "android", target_os = "linux"))]
-#[no_mangle]
-unsafe extern "C" fn copy_file_range(
-    fd_in: c_int,
-    off_in: *mut i64,
-    fd_out: c_int,
-    off_out: *mut i64,
-    len: usize,
-    flags: c_uint,
-) -> isize {
-    libc!(libc::copy_file_range(
-        fd_in, off_in, fd_out, off_out, len, flags
-    ));
-
-    if fd_in == -1 || fd_out == -1 {
-        set_errno(Errno(libc::EBADF));
-        return -1;
-    }
-    assert_eq!(flags, 0);
-    let off_in = if off_in.is_null() {
-        None
-    } else {
-        Some(&mut *off_in.cast::<u64>())
-    };
-    let off_out = if off_out.is_null() {
-        None
-    } else {
-        Some(&mut *off_out.cast::<u64>())
-    };
-    match convert_res(rustix::fs::copy_file_range(
-        &BorrowedFd::borrow_raw(fd_in),
-        off_in,
-        &BorrowedFd::borrow_raw(fd_out),
-        off_out,
-        len as u64,
-    )) {
-        Some(n) => n as _,
-        None => -1,
-    }
-}
-
-#[cfg(not(target_os = "wasi"))]
-#[no_mangle]
-unsafe extern "C" fn chmod(pathname: *const c_char, mode: c_uint) -> c_int {
-    libc!(libc::chmod(pathname, mode));
-
-    let mode = Mode::from_bits((mode & !libc::S_IFMT) as _).unwrap();
-    match convert_res(rustix::fs::chmodat(
-        &cwd(),
-        ZStr::from_ptr(pathname.cast()),
-        mode,
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[cfg(not(target_os = "wasi"))]
-#[no_mangle]
-unsafe extern "C" fn fchmod(fd: c_int, mode: c_uint) -> c_int {
-    libc!(libc::fchmod(fd, mode));
-
-    let mode = Mode::from_bits((mode & !libc::S_IFMT) as _).unwrap();
-    match convert_res(rustix::fs::fchmod(&BorrowedFd::borrow_raw(fd), mode)) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn linkat(
-    olddirfd: c_int,
-    oldpath: *const c_char,
-    newdirfd: c_int,
-    newpath: *const c_char,
-    flags: c_int,
-) -> c_int {
-    libc!(libc::linkat(olddirfd, oldpath, newdirfd, newpath, flags));
-
-    let flags = AtFlags::from_bits(flags as _).unwrap();
-    match convert_res(rustix::fs::linkat(
-        &BorrowedFd::borrow_raw(olddirfd),
-        ZStr::from_ptr(oldpath.cast()),
-        &BorrowedFd::borrow_raw(newdirfd),
-        ZStr::from_ptr(newpath.cast()),
-        flags,
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn symlink(target: *const c_char, linkpath: *const c_char) -> c_int {
-    libc!(libc::symlink(target, linkpath));
-
-    match convert_res(rustix::fs::symlinkat(
-        ZStr::from_ptr(target.cast()),
-        &cwd(),
-        ZStr::from_ptr(linkpath.cast()),
-    )) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[no_mangle]
-unsafe extern "C" fn chown() {
-    //libc!(libc::chown());
-    unimplemented!("chown")
-}
-
-#[no_mangle]
-unsafe extern "C" fn lchown() {
-    //libc!(libc::lchown());
-    unimplemented!("lchown")
-}
-
-#[no_mangle]
-unsafe extern "C" fn fchown() {
-    //libc!(libc::fchown());
-    unimplemented!("fchown")
 }
 
 // net
@@ -1649,14 +976,6 @@ unsafe extern "C" fn poll(fds: *mut rustix::io::PollFd, nfds: c_ulong, timeout: 
     }
 }
 
-#[no_mangle]
-unsafe extern "C" fn close(fd: c_int) -> c_int {
-    libc!(libc::close(fd));
-
-    rustix::io::close(fd);
-    0
-}
-
 #[cfg(not(target_os = "wasi"))]
 #[no_mangle]
 unsafe extern "C" fn dup2(fd: c_int, to: c_int) -> c_int {
@@ -2305,7 +1624,8 @@ unsafe extern "C" fn dlsym(handle: *mut c_void, symbol: *const c_char) -> *mut c
         // functions it asks for.
         #[cfg(any(target_os = "android", target_os = "linux"))]
         if ZStr::from_ptr(symbol.cast()).to_bytes() == b"statx" {
-            return statx as *mut c_void;
+            todo!()
+            // return statx as *mut c_void;
         }
         #[cfg(any(target_os = "android", target_os = "linux"))]
         if ZStr::from_ptr(symbol.cast()).to_bytes() == b"getrandom" {
@@ -2313,7 +1633,8 @@ unsafe extern "C" fn dlsym(handle: *mut c_void, symbol: *const c_char) -> *mut c
         }
         #[cfg(any(target_os = "android", target_os = "linux"))]
         if ZStr::from_ptr(symbol.cast()).to_bytes() == b"copy_file_range" {
-            return copy_file_range as *mut c_void;
+            // return copy_file_range as *mut c_void;
+            todo!()
         }
         #[cfg(any(target_os = "android", target_os = "linux"))]
         if ZStr::from_ptr(symbol.cast()).to_bytes() == b"clone3" {
